Speed up in monitoring up and down trends
You should notice a huge improvement of the monitoring speed of up and down trends. The previous version was more moderate. It's necessary that you uninstall a previous installed xDrip app and install the .apk file from the main directory (xDrip-Experimental-LimiTTer.apk). This version has a constant slope = 1.0. Maybe you have to enable "install from unknown source option" in your Android, because it's an unsigned .apk file. It's only a temporary app ... @AdrianLxM is working to include LimiTTer in xDrip. When finished the download will be from his repo. Thanks for your work!