Changed fetching method of shownGlucose
Problem of the previous one was that there were noisy values when glucose is stable.
This was because if the LimiTTer has detected that glucose is falling, it has taken the lowest
value of all trend values - and the highest if rising. It looked like a "ping-pong" trend. Now 
it only takes and shows the recent trend value. If this is not valid, it will taken the next 
one and if not again the next-next one. If this value  is still not valid (very corrupt scan), it will
then taken and shown the average of all valid trend values.